requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
Flask == 2.2.2
Flask-Caching == 2.0.1
Flask-Mail == 0.9.1
ibmcloudant == 0.3.0
ipython == 8.6.0
joblib == 1.2.0
matplotlib == 3.6.2
numpy == 1.23.4
opencv-contrib-python == 4.6.0.66
pandas == 1.5.1
playsound == 1.3.0
psutil == 5.9.3
python-dotenv == 0.21.0
pyyaml == 6.0
requests == 2.28.1
scipy == 1.9.3
seaborn == 0.12.1
sendgrid == 6.9.7
torch == 1.13.0
torchvision == 0.14.0
torchaudio == 0.13.0
tqdm == 4.64.1
Werkzeug == 2.2.2